refactor: separate package for converters
