+++ app/src/main/java/ai/elimu/analytics/db/RoomDb.java
@@ -22,8 +22,9 @@ import ai.elimu.analytics.dao.StoryBookLearningEventDao;
 import ai.elimu.analytics.dao.VideoLearningEventDao;
 import ai.elimu.analytics.dao.WordAssessmentEventDao;
 import ai.elimu.analytics.dao.WordLearningEventDao;
+import ai.elimu.analytics.db.converter.Converters;
 import ai.elimu.analytics.db.migration.AutoMigrationSpecFrom22To23;
-import ai.elimu.analytics.db.util.StringListConverter;
+import ai.elimu.analytics.db.converter.StringListConverter;
 import ai.elimu.analytics.entity.LetterSoundAssessmentEvent;
 import ai.elimu.analytics.entity.LetterSoundLearningEvent;
 import ai.elimu.analytics.entity.NumberAssessmentEvent;
